requirements.txt
greenlet==1.0.0
mysql-connector-python==8.0.23
protobuf==3.15.6
six==1.15.0
SQLAlchemy==1.4.3
